requirements.txt
# Core
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# TDA Engine
giotto-tda>=0.6.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0

# Web API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Visualization
plotly>=5.18.0

# Templating
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
